Enable nullable and fix all warnings in the project
Enable nullable.
All warnings that results from this also those who was there prior to
this pr is fixed.
Also fixes random bugs and other problems.

Namespaces has been fixed as well.

A lot of stuff is set after the constructor is run due to this being
unity. So for those we silence the null warning.
All references set in the editor are also null silenced.