Optimize Visible Tiles
Before it was a dictionary that was constantly recreated causing memory
allocations.
Now it is a fixed-size multi-dimensional array.

It keeps a generation, which is incremented by ResetRobotVisibility this
invalidates the current generation.
This makes it so we don't have to clear the array.

This completely removes it from the profiler.