Fix vertex cloning
Before when we cloned a vertex we kept all neighbors as the original
references to the original vertices. This was broken as after the robot
visited the first vertex it would go on to the global vertices and thus
share knowledge with other robots.

Now we clone a whole PatrollingMap. This makes it so that we can clone
all vertices in the map, changing the references to the neighbors to the
cloned vertices.

This uncovered some bugs that depended on that the algorithms would
visit the target vertex (which was form the global vertices) for
visualizing.

Also as a bonus made patrolling map generator twice as fast and reduce
code duplication in patrolling algorithms.