Fix collisionMap (all SimulationMaps)
Before it did some strange things with SquareGrid that was really broken
as the map did not align to CoarseMap tiles which made many places
impossible for the robot to navigate to.

Now it has been fixed by bypassing the squaregrid stuff and just using
the Tile[,] map to create the collision map.

Fixes a bug in cave map where the PatrollingMap threw an exception due
to Astar being unable to path to the waypoint.

Also when the robot collides with something it switches from using the
precomupted astar paths to the robotcontroller path and move to
functionality to hopefully navigate around the obstacle.

We now calculate paths both ways from a waypoint to its neighbor and
back.
This sometime creates two different paths (of the same length) which
helps with robots colliding with each other.

Co-Authored-By: Casper-NS <37834568+Casper-NS@users.noreply.github.com>