Fix Cognitive Coordinated
Right now it does nothing with the vertices received from other robots.
This fixes that by using vertex ids instead of vertex objects.

This uncovered a bug where it tried to start from the nearest waypoint
causing pathfinding from and to the same waypoint, which is not in the
dictionary of paths causing an exception.