Implement new Line-of-sight algorithm
This algorithm is a grid traversal algorithm.
It will not skip over places where the line barely touches.
This means that this algorithm generate visibility where the squares
sides are always touching, there are no floating "islands" or
connections only by corners.

It is implemented with Burst which compiles a subset of C# into highly
optimized and vectorized code.
This greatly improves the startup time.
On my pc it takes just 4-5 seconds to run visibility calculation for
100x100 maps.

Also implements a new Bitmap class that is a 2d boolean array stored as
an array of 64 bit integers.
This decreases the memory used for 2d maps.

The PatrollingAlgorithm path to next vertex stuff has been fixed.