Implement new Line-of-sight algorithm
This algorithm is a grid traversal algorithm.
It will not skip over places where the line barely touches.
This means that this algorithm generate visibility where the squares
sides are always touching, there are no floating "islands" or
connections only by corners.

Also implements a new Bitmap class that is a 2d boolean array stored as
an array of 64 bit integers.
This decreases the memory used for 2d maps.

Lastly an optional range parameter can be given to the los algorithm to
limit the range of a waypoint.
This generates a lot more waypoints, but is more realistic, as the
robots does not have infinite visibility.