Make PatrollingAlgorithm Component based
Components have two methods `PreUpdate` and `PostUpdate`.

`PreUpdate` is called before the algorithm's `Update` method is called.
Here the component do stuff such as move the robot about.
It can also wait for certain conditions such as a robot status or
waiting for x ticks.
Likewise `PostUpdate` is run after the algorithm's `Update` method is
called.

`PreUpdate` and `PostUpdate` are enumerables so that we can implement
statemachines without having to keep track of the machine's state
manually.

They return a wrapper around `WaitForCondition` with an additional bool
`ShouldContinue`.
If `ShouldContinue` is true the next component in the chain is run, and
eventually `Update` is run.

If `ShouldContinue` is false the chain stops executing there and next
tick `PreUpdate` is called again. Here `Update` and `PostUpdate` is not
called.
This allows the components to do exceptional things such as collision
avoidance, while not being trampled over by the next components.