Add Startup Component
This component does some computation on robot 0 and sends it to all
other robots in range. The result of the computation can be accessed as
well as all the robot IDs that were discovered doing the startup
process.

Robots not in range of robot 0 will disable themselves and run no
further code.