Make some stuff forbidden from being accessed in an Algorithm
This adds a new analyzer that stops the user from accessing certain
things from algorithms and components.
This is to stop the user from doing wildly unrealistic or unsupported
things.

A new attribute `ForbiddenKnowledgeAttribute` has been added to mark
methods, properties, and constructors as forbidden knowledge, e.g. must
not be referenced in an algorithm.

I have made a new analyzer that analyzes all call instructions in the
project.
It is a bit complicated, but it is not too bad.