Multi-Instance simulation
When building the run.sh and run-headless.sh scripts allow the user to
run multiple instances of unity at once, each taking a part of the
workload.

This is useful when we have many scenarios as this allows us to
"multithread" the simulations.

Run the scripts by `run{,-headless}.sh <instances>`.

Also allow to build and run for Linux server target.