Cache waypoint generator
We currently only have one waypoint generator, but this caching
mechanism can be used for any waypoint generator.

It only stores the hash of the bitmap (and extra data), and a hashset of
vertex positions.
This means that we can basically store infinitely many things in our
cache, as it uses almost no memory.
Furthermore, as the cache is entirely in memory, no issues with old
invalid caches will happen.