Optimize creating a new simulation
Before we did a lot of unecessary work.
I have optimized it to allocate a lot less memory on start.
Communication zones are now lazy so they are not slowing down everything
if they are not used.
A full garbage collection is run after a simulation has been created to
hopefully reduce memory usage.