More optimizations
Optimize Visibility calculations by only tracing to the border of the
map. This is inaccurate though, but should be good enough for the things
we use visibility calculations for.

Multithread generating paths between vertices on patrolling map. This is
especially noticable when using everyone connected to everyone and or on
big maps.

Use a SortedSet in Greedy Most Visibility Waypoint Connector to avoid
having to sort every iteration. (about 33% performance improvement).

Only calculate Distance Matrix once. This is extremely expensive and
doing this multiple times was extremely slow.

Other small optimizations.